tests/test_requirements.txt
pytest
pytest-cov
codecov
click
mkdocs-material